Safari-611.1.10/LayoutTests/http/tests/uri/escaped-entity.html

<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=big5">
</head>
<body>

<script>
    if (window.testRunner)
        testRunner.dumpAsText();
</script>

<p>Test for <a href='http://bugs.webkit.org/show_bug.cgi?id=15119'>bug 15119</a>:
Unrepresentable characters in a URL's character set should be converted to
escaped entities. We use the character U+06DE (ARABIC START OF RUB EL HIZB) which does not exist in
Big-5.</p>

<p>Note that this exact page won't work in IE or Firefox. Firefox seems to
always use UTF-8 for local files, and IE actually preserves the Unicode in the
URL when we get it from JS, so we don't know what would get sent over the wire.
However, both browsers will send <tt>%26%231758%3B</tt> over HTTP for the
query.</p>

"<script id=scr1 src="intercept/print/script.js?&#x6DE;"></script>" (no target charset specified, should be Big5)<br>
"<script id=scr2 charset="big5" src="intercept/print/script.js?&#x6DE;"></script>" (Big5 specified)<br>

Show the source attribute of the scripts.<br>
<script>
 try {
  document.write('"' + document.scripts[1].src + '"<br>');
  document.write('"' + document.scripts[2].src + '"<br>');
 } catch (ex) {
  document.write('"' + document.getElementById("scr1").src + '"<br>');
  document.write('"' + document.getElementById("scr2").src + '"<br>');
 }
</script>

</body>
</html>
