third_party/blink/web_tests/http/tests/uri/escaped-entity-expected.txt

Test for bug 15119: Unrepresentable characters in a URL's character set should be converted to escaped entities. We use the character U+06DE (ARABIC START OF RUB EL HIZB) which does not exist in Big-5.

Note that this exact page won't work in IE or Firefox. Firefox seems to always use UTF-8 for local files, and IE actually preserves the Unicode in the URL when we get it from JS, so we don't know what would get sent over the wire. However, both browsers will send %26%231758%3B over HTTP for the query.

"/uri/intercept/print/script.js?%26%231758%3B" (no target charset specified, should be Big5)
"/uri/intercept/print/script.js?%26%231758%3B" (Big5 specified)
Show the source attribute of the scripts.
"http://127.0.0.1:8000/uri/intercept/print/script.js?%26%231758%3B"
"http://127.0.0.1:8000/uri/intercept/print/script.js?%26%231758%3B"
